Refactor into a Lite version w/ just Start, Stop and Save messages. Released as v3.0.0
