--- NuGet.config
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <packageSources>
-    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" protocolVersion="3" />
-    <add key="BepInEx" value="https://nuget.bepinex.dev/v3/index.json" />
-  </packageSources>
-</configuration>
