+++ src/DiscordConnector/DiscordConnectorLite.csproj
@@ -1,8 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net48</TargetFramework>
-    <RootNamespace>DiscordConnector</RootNamespace>
+    <RootNamespace>DiscordConnectorLite</RootNamespace>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <LangVersion>latest</LangVersion>
     <OutputType>Library</OutputType>
@@ -16,19 +16,6 @@
     <DestinationDir Condition=" '$(NoBotSupport)' == '' ">$(ProjectDir)bin/DiscordConnector/</DestinationDir>
     <DestinationDir Condition=" '$(NoBotSupport)' == '1' ">$(ProjectDir)bin/DiscordConnector-NoBotSupport/</DestinationDir>
   </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="BepInEx.Analyzers" Version="1.0.4" PrivateAssets="all" />
-    <PackageReference Include="BepInEx.Core" Version="5.*" />
-    <PackageReference Include="LiteDB" Version="5.*" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.*" />
-    <PackageReference Include="UnityEngine.Modules" Version="2019.4.24" IncludeAssets="compile" />
-    <PackageReference Include="Valheim.GameLibs" Version="*-*" />
-  </ItemGroup>
-  
-  <ItemGroup Condition="'$(TargetFramework.TrimEnd(`0123456789`))' == 'net'">
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.2" PrivateAssets="all" />
-  </ItemGroup>
 
    <Target Name="CreatePluginDir" AfterTargets="PrepareForBuild">
     <MakeDir Directories="$(DestinationDir)" Condition="!Exists('$(DestinationDir)')" />
@@ -64,4 +51,33 @@
     <Delete Files="$(ProjectDir)bin/DiscordConnector.zip" Condition="!Exists('$(ProjectDir)bin/bin/DiscordConnector.zip')" />
     <ZipDirectory SourceDirectory="$(DestinationDir)" DestinationFile="$(ProjectDir)bin/DiscordConnector.zip" />
   </Target>
+
+  <ItemGroup>
+    <PackageReference Include="BepInEx.Core" Version="[5.4.21]">
+      <PrivateAssets>all</PrivateAssets>
+      <ExcludeAssets>runtime</ExcludeAssets>
+    </PackageReference>
+    <PackageReference Include="Digitalroot.Valheim.Build.Targets" Version="1.0.21">
+      <PrivateAssets>all</PrivateAssets>
+      <ExcludeAssets>runtime</ExcludeAssets>
+    </PackageReference>
+    <PackageReference Include="Digitalroot.References.Unity" Version="2022.3.12" >
+      <PrivateAssets>all</PrivateAssets>
+      <ExcludeAssets>runtime</ExcludeAssets>
+      <IncludeAssets>compile; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Digitalroot.Valheim.Common.References" Version="0.217.38" >
+      <PrivateAssets>all</PrivateAssets>
+      <ExcludeAssets>runtime</ExcludeAssets>
+      <IncludeAssets>compile; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    <PackageReference Include="Fody" Version="[6.6.0]">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="ILMerge.Fody" Version="[1.22.0]">
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
+  </ItemGroup>
 </Project>
