+++ src/DiscordConnector/Metadata/CHANGELOG.md
