+++ src/DiscordConnector/Metadata/DiscordConnector-Nexus.readme
