+++ src/DiscordConnector/Metadata/README.md
