+++ src/DiscordConnector/Metadata/icon.pdn
