+++ src/DiscordConnector/Metadata/icon.png
