+++ src/DiscordConnector/Metadata/manifest.json
