+++ src/DiscordConnector/PluginInfo.cs
@@ -1,5 +1,5 @@
 ﻿using System.Reflection;
-using DiscordConnector;
+using DiscordConnectorLite;
 
 #region Assembly attributes
 /*
@@ -8,10 +8,10 @@ using DiscordConnector;
 [assembly: AssemblyVersion(PluginInfo.PLUGIN_VERSION)]
 [assembly: AssemblyTitle(PluginInfo.PLUGIN_NAME + " (" + PluginInfo.PLUGIN_ID + ")")]
 [assembly: AssemblyProduct(PluginInfo.PLUGIN_NAME)]
-[assembly: AssemblyCopyright("© 2021 " + PluginInfo.PLUGIN_AUTHOR + " Repository at " + PluginInfo.PLUGIN_REPO_SHORT)]
+[assembly: AssemblyCopyright("© 2021 - 2024" + PluginInfo.PLUGIN_AUTHOR + " Repository at " + PluginInfo.PLUGIN_REPO_SHORT)]
 #endregion
 
-namespace DiscordConnector
+namespace DiscordConnectorLite
 {
     internal static class PluginInfo
     {
@@ -19,12 +19,12 @@ namespace DiscordConnector
         public const string PLUGIN_ID = "games.nwest.valheim.discordconnector-nobotsupport";
         public const string PLUGIN_NAME = "Valheim Discord Connector (No Discord Bot Support)";
 #else
-        public const string PLUGIN_ID = "games.nwest.valheim.discordconnector";
-        public const string PLUGIN_NAME = "Valheim Discord Connector";
+        public const string PLUGIN_ID = "games.nwest.valheim.discordconnector.lite";
+        public const string PLUGIN_NAME = "Valheim Discord Connector Lite";
 #endif
 
-        public const string PLUGIN_VERSION = "1.4.3";
-        public const string PLUGIN_REPO_SHORT = "github: nwesterhausen/valheim-discordconnector";
-        public const string PLUGIN_AUTHOR = "Nicholas Westerhausen";
+        public const string PLUGIN_VERSION = "3.0.0";
+        public const string PLUGIN_REPO_SHORT = "github: Digitalroot-Valheim/nwesterhausen-valheim-discordconnector";
+        public const string PLUGIN_AUTHOR = "Nicholas Westerhausen and Digitalroot";
     }
 }
